+++ CHANGELOG.md
@@ -42,6 +42,7 @@ This repo attempts to follow [semantic versioning](https://semver.org/).
 - `OccurrenceStream.isOccurrenceStream()`
 - ensure `DateAdapter#date` is immutable
 - don't include `undefined` properties in `DateAdapter#toJSON()`
+- ensure `DateAdapter#generators` is propogated to results
 
 ## 0.10.0 BREAKING (2019/4/15)